changed HomeFeed flow
