+++ likeminds-chat-reactnative-integration/ChatSX/context/GroupFeedContextProvider.tsx
@@ -35,11 +35,11 @@ import { useIsFocused, useNavigation } from "@react-navigation/native";
 import { StackNavigationProp } from "@react-navigation/stack";
 import { SHOW_LIST_REGEX } from "../commonFuctions";
 
-export interface HomeFeedContextProps {
+export interface GroupFeedContextProps {
   children?: ReactNode;
 }
 
-export interface HomeFeedContext {
+export interface GroupFeedContext {
   isLoading: any;
   shimmerIsLoading: any;
   invitePage: any;
@@ -48,13 +48,6 @@ export interface HomeFeedContext {
   groupFeedChatrooms: any;
   unseenCount: any;
   totalCount: any;
-  isLoadingDM: any;
-  shimmerIsLoadingDM: any;
-  showDM: any;
-  showList: any;
-  FCMTokenDM: any;
-  chatroomsDM: any;
-  dmFeedChatrooms: any;
 
   setIsLoading: Dispatch<SetStateAction<boolean>>;
   setShimmerIsLoading: Dispatch<SetStateAction<boolean>>;
@@ -65,40 +58,31 @@ export interface HomeFeedContext {
   getAppConfig: any;
   handleLoadMore: any;
   renderFooter: any;
-  setIsLoadingDM: Dispatch<SetStateAction<boolean>>;
-  setShimmerIsLoadingDM: Dispatch<SetStateAction<boolean>>;
-  setShowDM: Dispatch<SetStateAction<boolean>>;
-  setShowList: any;
-  setFCMTokenDM: any;
-  getChatroomFromLocalDBDM: any;
-  getAppConfigDM: any;
-  updateDataDM: any;
-  loadDataDM: any;
-  handleLoadMoreDM: any;
-  renderFooterDM: any;
 }
 
-const HomeFeedContext = createContext<HomeFeedContext | undefined>(undefined);
+const GroupFeedContext = createContext<GroupFeedContext | undefined>(undefined);
 
-export const useHomeFeedContext = () => {
-  const context = useContext(HomeFeedContext);
+export const useGroupFeedContext = () => {
+  const context = useContext(GroupFeedContext);
   if (!context) {
     throw new Error(
-      "useHomeFeedContext must be used within an HomeFeedContextProvider"
+      "useGroupFeedContext must be used within an GroupFeedContextProvider"
     );
   }
   return context;
 };
 
-export const HomeFeedContextProvider = ({ children }: HomeFeedContextProps) => {
+export const GroupFeedContextProvider = ({
+  children,
+}: GroupFeedContextProps) => {
   const myClient = Client.myClient;
   const [isLoading, setIsLoading] = useState(false);
   const [shimmerIsLoading, setShimmerIsLoading] = useState(true);
   const [invitePage, setInvitePage] = useState(1);
   const [FCMToken, setFCMToken] = useState("");
   const isFocused = useIsFocused();
-  const dispatch = useAppDispatch();
   const navigation = useNavigation<StackNavigationProp<any>>();
+  const dispatch = useAppDispatch();
 
   let {
     myChatrooms,
@@ -284,186 +268,13 @@ export const HomeFeedContextProvider = ({ children }: HomeFeedContextProps) => {
 
   const renderFooter = () => {
     return isLoading ? (
-      <View style={{ paddingVertical: Layout.normalize(20) }}>
+      <View style={{ paddingVertical: 20 }}>
         <ActivityIndicator size="large" color={STYLES.$COLORS.SECONDARY} />
       </View>
     ) : null;
   };
 
-  const [isLoadingDM, setIsLoadingDM] = useState(false);
-  const [shimmerIsLoadingDM, setShimmerIsLoadingDM] = useState(true);
-  const [showDM, setShowDM] = useState(false);
-  const [showList, setShowList] = useState<any>(null);
-  const [FCMTokenDM, setFCMTokenDM] = useState("");
-
-  const { myDMChatrooms, dmPage, dmFeedChatrooms } = useAppSelector(
-    (state) => state.homefeed
-  );
-
-  const chatroomsDM = [...myDMChatrooms];
-  const INITIAL_SYNC_PAGE_DM = 1;
-  let startTimeDM = 0;
-  let endTimeDM = 0;
-
-  async function fetchDataDM() {
-    if (community?.id) {
-      const payload = {
-        page: 1,
-      };
-      const apiRes = await myClient?.checkDMStatus({
-        requestFrom: "dm_feed_v2",
-      });
-      const response = apiRes?.data;
-      if (response) {
-        const routeURL = response?.cta;
-        const hasShowList = SHOW_LIST_REGEX.test(routeURL);
-        if (hasShowList) {
-          const showListValue = routeURL.match(SHOW_LIST_REGEX)[1];
-          setShowList(showListValue);
-        }
-        setShowDM(response?.showDm);
-      }
-    }
-  }
-
-  useEffect(() => {
-    if (isFocused) {
-      LMChatAnalytics.track(
-        Events.DM_FEED_OPENED,
-        new Map<string, string>([[Keys.SOURCE, Sources.HOME_FEED]])
-      );
-    }
-  }, [isFocused]);
-
-  useLayoutEffect(() => {
-    fetchData();
-  }, [navigation, community]);
-
-  useEffect(() => {
-    const token = async () => {
-      const isPermissionEnabled = await requestUserPermission();
-      if (isPermissionEnabled) {
-        const fcmToken = await fetchFCMToken();
-        if (fcmToken) {
-          setFCMToken(fcmToken);
-        }
-      }
-    };
-    token();
-  }, []);
-
-  // Fetching already existing chatrooms from Realm
-  const getChatroomFromLocalDBDM = async () => {
-    const existingChatrooms: any = await myClient?.getFilteredChatrooms(true);
-    if (!!existingChatrooms && existingChatrooms.length !== 0) {
-      setShimmerIsLoading(false);
-      dispatch({
-        type: SET_INITIAL_DMFEED_CHATROOM,
-        body: { dmFeedChatrooms: existingChatrooms },
-      });
-    }
-  };
-
-  const getAppConfigDM = async () => {
-    const appConfig = await myClient?.getAppConfig();
-    if (appConfig?.isDmFeedChatroomsSynced === undefined) {
-      startTimeDM = Date.now() / 1000;
-      setTimeout(() => {
-        myClient?.initiateAppConfig();
-        myClient?.setAppConfig(true);
-      }, 200);
-    } else {
-      setShimmerIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    const query = ref(db, `/community/${community?.id}`);
-    return onValue(query, (snapshot) => {
-      if (snapshot.exists()) {
-        if (!user?.sdkClientInfo?.community) {
-          return;
-        }
-        if (isFocused) {
-          paginatedSyncAPI(INITIAL_SYNC_PAGE_DM, user, true);
-          setShimmerIsLoading(false);
-          setTimeout(() => {
-            getChatroomFromLocalDB();
-          }, 500);
-        }
-      }
-    });
-  }, [user, isFocused]);
-
-  useEffect(() => {
-    const initiate = async () => {
-      await getAppConfig();
-      if (!user?.sdkClientInfo?.community) {
-        return;
-      }
-      await paginatedSyncAPI(INITIAL_SYNC_PAGE_DM, user, true);
-      if (shimmerIsLoading == true && isFocused) {
-        endTimeDM = Date.now() / 1000;
-        LMChatAnalytics.track(
-          Events.SYNC_COMPLETE,
-          new Map<string, string>([
-            [Keys.SYNC_COMPLETE, true?.toString()],
-            [Keys.TIME_TAKEN, (endTimeDM - startTimeDM)?.toString()],
-          ])
-        );
-      }
-      setShimmerIsLoading(false);
-      setTimeout(() => {
-        getChatroomFromLocalDB();
-      }, 500);
-    };
-    initiate();
-  }, [user, isFocused, shimmerIsLoading, startTimeDM]);
-
-  //function calls updateDMFeedData action to update myDMChatrooms array with the new data.
-  async function updateDataDM(newPage: number) {
-    const payload = {
-      page: newPage,
-    };
-    const response = await dispatch(updateDMFeedData(payload, false) as any);
-    return response;
-  }
-
-  // function shows loader in between calling the API and getting the response
-  const loadDataDM = async (newPage: number) => {
-    setIsLoading(true);
-    setTimeout(async () => {
-      const res: any = await updateData(newPage);
-      if (res) {
-        setIsLoading(false);
-      }
-    }, 1500);
-  };
-
-  //function checks the pagination logic, if it verifies the condition then call loadData
-  const handleLoadMoreDM = async () => {
-    if (!isLoading) {
-      if (
-        myDMChatrooms?.length > 0 &&
-        myDMChatrooms?.length % 10 === 0 &&
-        myDMChatrooms?.length === 10 * dmPage
-      ) {
-        const newPage = dmPage + 1;
-        dispatch({ type: SET_DM_PAGE, body: newPage });
-        loadData(newPage);
-      }
-    }
-  };
-
-  const renderFooterDM = () => {
-    return isLoading ? (
-      <View style={{ paddingVertical: Layout.normalize(20) }}>
-        <ActivityIndicator size="large" color={STYLES.$COLORS.SECONDARY} />
-      </View>
-    ) : null;
-  };
-
-  const contextValues: HomeFeedContext = {
+  const contextValues: GroupFeedContext = {
     isLoading,
     shimmerIsLoading,
     invitePage,
@@ -472,13 +283,6 @@ export const HomeFeedContextProvider = ({ children }: HomeFeedContextProps) => {
     groupFeedChatrooms,
     unseenCount,
     totalCount,
-    isLoadingDM,
-    shimmerIsLoadingDM,
-    showDM,
-    showList,
-    FCMTokenDM,
-    chatroomsDM,
-    dmFeedChatrooms,
 
     setIsLoading,
     setShimmerIsLoading,
@@ -489,22 +293,11 @@ export const HomeFeedContextProvider = ({ children }: HomeFeedContextProps) => {
     getAppConfig,
     handleLoadMore,
     renderFooter,
-    setIsLoadingDM,
-    setShimmerIsLoadingDM,
-    setShowDM,
-    setShowList,
-    setFCMTokenDM,
-    getChatroomFromLocalDBDM,
-    getAppConfigDM,
-    updateDataDM,
-    loadDataDM,
-    handleLoadMoreDM,
-    renderFooterDM,
   };
 
   return (
-    <HomeFeedContext.Provider value={contextValues}>
+    <GroupFeedContext.Provider value={contextValues}>
       {children}
-    </HomeFeedContext.Provider>
+    </GroupFeedContext.Provider>
   );
 };
