+++ likeminds-chat-reactnative-integration/ChatSX/context/InputBoxContext.tsx
@@ -26,18 +26,6 @@ import {
 import AWS, { CognitoIdentityCredentials, S3 } from "aws-sdk";
 import { POOL_ID, REGION } from "../awsExports";
 import {
-  useAnimatedStyle,
-  useSharedValue,
-  withRepeat,
-  withSpring,
-  withTiming,
-} from "react-native-reanimated";
-import {
-  Gesture,
-  GestureUpdateEvent,
-  PanGestureHandlerEventPayload,
-} from "react-native-gesture-handler";
-import {
   atLeastAndroid13,
   detectMentions,
   extractPathfromRouteQuery,
@@ -146,13 +134,6 @@ export interface InputBoxContextProps {
 export interface InputBoxContext {
   isVoiceNoteIconPress: any;
   hideDMSentAlert: any;
-  lockAnimatedStyles: any;
-  upChevronAnimatedStyles: any;
-  longPressGesture: any;
-  onUpdatePanGesture: any;
-  onEndPanGesture: any;
-  panGesture: any;
-  panStyle: any;
   handleVideoThumbnail: any;
   handleImageAndVideoUpload: any;
   selectGallery: any;
@@ -173,14 +154,6 @@ export interface InputBoxContext {
   detectLinkPreview: any;
   handleInputChange: any;
   onEdit: any;
-  startRecord: any;
-  stopRecord: any;
-  handleStopRecord: any;
-  clearVoiceRecord: any;
-  startPlay: any;
-  stopPlay: any;
-  onPausePlay: any;
-  onResumePlay: any;
   askPermission: any;
   setModalVisible: any;
   chatroomType: any;
@@ -199,7 +172,6 @@ export interface InputBoxContext {
   groupTags: any;
   taggedUserName: any;
   setMessage: any;
-  setFormattedConversation: any;
   setUserTaggingList: any;
   setGroupTags: any;
   setIsUserTagging: any;
@@ -214,28 +186,22 @@ export interface InputBoxContext {
   setIsEditable: any;
   isDoc: any;
   isGif: any;
-  isDeleteAnimation: any;
   voiceNotes: any;
-  isVoiceResult: any;
-  micIconOpacity: any;
   isRecordingLocked: any;
-  isVoiceNotePlaying: any;
-  voiceNotesPlayer: any;
-  voiceNotesLink: any;
   chatRequestState: any;
   inputHeight: any;
   setInputHeight: any;
   myRef: any;
   MAX_LENGTH: any;
   isPrivateMember: any;
-  isVoiceNoteRecording: any;
   isRecordingPermission: any;
-  composedGesture: any;
   setIsVoiceNoteIconPress: any;
   modalVisible: any;
   navigation: any;
   chatroomID: any;
   conversations: any;
+  setIsRecordingLocked: any;
+  setVoiceNotes: any;
 }
 
 const InputBoxContext = createContext<InputBoxContext | undefined>(undefined);
@@ -273,19 +239,11 @@ export const InputBoxContextProvider = ({
 }: InputBoxContextProps) => {
   const myClient = Client.myClient;
   const inputBoxStyles = STYLES.$INPUT_BOX_STYLE;
-  const audioRecorderPlayerAttachment = new AudioRecorderPlayer();
 
   const [isKeyBoardFocused, setIsKeyBoardFocused] = useState(false);
   const [message, setMessage] = useState(previousMessage);
-  const [formattedConversation, setFormattedConversation] =
-    useState(previousMessage);
   const [inputHeight, setInputHeight] = useState(25);
-  const [showEmoji, setShowEmoji] = useState(false);
   const [modalVisible, setModalVisible] = useState(false);
-  const [pollModalVisible, setPollModalVisible] = useState(false);
-  const [progressText, setProgressText] = useState("");
-  const [isUploading, setIsUploading] = useState(false);
-  const [s3UploadResponse, setS3UploadResponse] = useState<any>();
   const [DMSentAlertModalVisible, setDMSentAlertModalVisible] = useState(false);
   const [debounceTimeout, setDebounceTimeout] = useState<any>(null);
   const [debounceLinkPreviewTimeout, setLinkPreviewDebounceTimeout] =
@@ -302,24 +260,11 @@ export const InputBoxContextProvider = ({
     recordTime: "",
     name: "",
   });
-  const [voiceNotesPlayer, setVoiceNotesPlayer] =
-    useState<VoiceNotesPlayerProps>({
-      currentPositionSec: 0,
-      currentDurationSec: 0,
-      playTime: "",
-      duration: "",
-    });
-  const [voiceNotesLink, setVoiceNotesLink] = useState("");
   const [isVoiceResult, setIsVoiceResult] = useState(false);
-  const [isVoiceNotePlaying, setIsVoiceNotePlaying] = useState(false);
-  const [isVoiceNoteRecording, setIsVoiceNoteRecording] = useState(false);
-  const [isDraggable, setIsDraggable] = useState(true);
-  const [isLongPressedState, setIsLongPressedState] = useState(false);
   const [isRecordingLocked, setIsRecordingLocked] = useState(false);
-  const [stopRecording, setStopRecording] = useState(false);
-  const [isDeleteAnimation, setIsDeleteAnimation] = useState(false);
   const [isRecordingPermission, setIsRecordingPermission] = useState(false);
   const [isVoiceNoteIconPress, setIsVoiceNoteIconPress] = useState(false);
+
   const { chatroomDBDetails }: any = useAppSelector((state) => state.chatroom);
 
   const [ogTagsState, setOgTagsState] = useState<any>({});
@@ -344,9 +289,7 @@ export const InputBoxContextProvider = ({
     conversations = [],
     selectedMessages = [],
   }: any = useAppSelector((state) => state.chatroom);
-  const { myChatrooms, user, community }: any = useAppSelector(
-    (state) => state.homefeed
-  );
+  const { user, community }: any = useAppSelector((state) => state.homefeed);
   const {
     chatroomDetails,
     isReply,
@@ -355,13 +298,9 @@ export const InputBoxContextProvider = ({
     fileSent,
   }: any = useAppSelector((state) => state.chatroom);
 
-  const { uploadingFilesMessages }: any = useAppSelector(
-    (state) => state.upload
-  );
   let isGroupTag = false;
 
   const dispatch = useAppDispatch();
-  const conversationArrayLength = conversations.length;
 
   AWS.config.update({
     region: REGION, // Replace with your AWS region, e.g., 'us-east-1'
@@ -371,182 +310,6 @@ export const InputBoxContextProvider = ({
   });
 
   const s3 = new S3();
-
-  // Animation
-
-  const pressed = useSharedValue(false);
-  const x = useSharedValue(0);
-  const y = useSharedValue(0);
-  const lockOffset = useSharedValue(4);
-  const upChevronOffset = useSharedValue(3);
-  const micIconOpacity = useSharedValue(1); // Initial opacity value
-  const isLongPressed = useSharedValue(false);
-
-  // lock icon animation styles
-  const lockAnimatedStyles = useAnimatedStyle(() => ({
-    transform: [{ translateY: lockOffset.value }],
-  }));
-
-  // up chevron animated styles
-  const upChevronAnimatedStyles = useAnimatedStyle(() => ({
-    transform: [{ translateY: upChevronOffset.value }],
-  }));
-
-  // recorder mic icon animation effect
-  useEffect(() => {
-    micIconOpacity.value = withRepeat(
-      withTiming(0, { duration: 800, easing: Easing.inOut(Easing.ease) }),
-      -1, // Infinite repetition (-1)
-      true // Reverse the animation direction after each iteration
-    );
-  }, []);
-
-  // lock icon animation useEffect
-  useEffect(() => {
-    lockOffset.value = withRepeat(
-      withTiming(-lockOffset.value, { duration: 800 }),
-      -1, // Infinite repetition (-1)
-      true // Reverse the animation direction after each iteration
-    );
-  }, []);
-
-  // up chevron animation useEffect
-  useEffect(() => {
-    upChevronOffset.value = withRepeat(
-      withTiming(-upChevronOffset.value, { duration: 400 }),
-      -1,
-      true
-    );
-  }, []);
-
-  // to hide delete animation
-  useEffect(() => {
-    if (isDeleteAnimation) {
-      setTimeout(() => {
-        setIsDeleteAnimation(false);
-        setIsVoiceResult(false);
-      }, 2200);
-    }
-  }, [isDeleteAnimation]);
-
-  // to stop the recorder
-  useEffect(() => {
-    setTimeout(async () => {
-      if (!isLongPressedState && isVoiceNoteRecording && !isRecordingLocked) {
-        await stopRecord();
-        setIsVoiceResult(true);
-      }
-    }, 300);
-  }, [isLongPressedState, isRecordingLocked]);
-
-  // to start Recorder
-  useEffect(() => {
-    if (isLongPressedState && !isVoiceNoteRecording) {
-      Vibration.vibrate(0.5 * 100);
-      startRecord();
-    }
-  }, [isLongPressedState]);
-
-  // long press gesture
-  const longPressGesture = Gesture.LongPress()
-    .runOnJS(true)
-    .minDuration(250)
-    .onStart(() => {
-      isLongPressed.value = true;
-      setIsLongPressedState(true);
-    });
-
-  // this method handles onUpdate callback of pan gesture
-  const onUpdatePanGesture = (
-    event: GestureUpdateEvent<PanGestureHandlerEventPayload>
-  ) => {
-    "worklet";
-    if (isLongPressed.value) {
-      if (Math.abs(x.value) >= 120) {
-        x.value = withSpring(0);
-        if (isDraggable) {
-          stopRecord();
-          setIsDraggable(false);
-          setIsDeleteAnimation(true);
-          clearVoiceRecord();
-          setIsDraggable(true);
-          isLongPressed.value = false;
-        }
-        pressed.value = false;
-        isLongPressed.value = false;
-      } else if (Math.abs(y.value) >= 100) {
-        y.value = withSpring(0);
-        if (isDraggable) {
-          setIsDraggable(false);
-          setIsDraggable(true);
-          setIsRecordingLocked(true);
-          setIsLongPressedState(false);
-          isLongPressed.value = false;
-        }
-      } else if (Math.abs(x.value) > 5) {
-        x.value = event.translationX;
-      } else if (Math.abs(y.value) > 5) {
-        y.value = event.translationY;
-      } else {
-        x.value = event.translationX;
-        y.value = event.translationY;
-      }
-    }
-  };
-
-  // this method handles onEnd callback of pan gesture
-  const onEndPanGesture = () => {
-    "worklet";
-    if (
-      (Math.abs(x.value) > 5 && Math.abs(x.value) < 120) ||
-      (Math.abs(y.value) > 5 && Math.abs(y.value) < 100)
-    ) {
-      setIsRecordingLocked(false);
-      handleStopRecord();
-    }
-    x.value = withSpring(0);
-    y.value = withSpring(0);
-    pressed.value = false;
-    isLongPressed.value = false;
-    setIsLongPressedState(false);
-  };
-
-  // draggle mic pan gesture on x-axis and y-axis
-  const panGesture = Gesture.Pan()
-    .runOnJS(true)
-    .enabled(isDraggable)
-    .onUpdate(onUpdatePanGesture)
-    .onEnd(onEndPanGesture)
-    .onFinalize(() => {
-      "worklet";
-      pressed.value = false;
-      isLongPressed.value = false;
-      setIsLongPressedState(false);
-      setIsDraggable(true);
-    })
-    .onTouchesCancelled(() => {
-      setIsDraggable(true);
-    })
-    .simultaneousWithExternalGesture(longPressGesture);
-
-  const composedGesture = Gesture.Simultaneous(longPressGesture, panGesture);
-
-  // draggle mic panGesture styles
-  const panStyle = useAnimatedStyle(() => {
-    return {
-      transform: [
-        {
-          translateX: x.value,
-        },
-        {
-          translateY: y.value,
-        },
-        { scale: withTiming(isLongPressed.value ? 1.5 : 1) },
-      ],
-    };
-  }, [x, y]);
-
-  // Animation stop
 
   // to ask audio recording permission
   useEffect(() => {
@@ -600,18 +363,9 @@ export const InputBoxContextProvider = ({
   useEffect(() => {
     if (!isUploadScreen) {
       setMessage("");
-      setFormattedConversation("");
       setInputHeight(25);
     }
   }, [fileSent]);
-
-  // this useEffect is to stop audio player when going out of chatroom, if any audio is running
-  useEffect(() => {
-    return () => {
-      stopRecord();
-      stopPlay();
-    };
-  }, []);
 
   // to clear message on ChatScreen InputBox when fileSent from UploadScreen
   useEffect(() => {
@@ -1637,11 +1391,9 @@ export const InputBoxContextProvider = ({
         });
       } else if (event.length < MAX_LENGTH) {
         setMessage(event);
-        setFormattedConversation(event);
       }
     } else {
       setMessage(event);
-      setFormattedConversation(event);
 
       // chatroomType === ChatroomType.DMCHATROOM (if DM don't detect and show user tags)
       const newMentions =
@@ -1790,191 +1542,6 @@ export const InputBoxContextProvider = ({
     );
   };
 
-  // Audio and play section
-
-  // to stop recorder after 15 min
-  useEffect(() => {
-    (async function stopRecorder() {
-      if (isVoiceNoteRecording) {
-        await handleStopRecord();
-      }
-    })();
-  }, [stopRecording]);
-
-  // to start audio recording
-  const startRecord = async () => {
-    if (!isVoiceNoteRecording) {
-      const audioSet = generateAudioSet();
-
-      const name = generateVoiceNoteName();
-      const path =
-        Platform.OS === "android"
-          ? `${ReactNativeBlobUtil.fs.dirs.CacheDir}/${name}.mp3`
-          : `${name}.m4a`;
-
-      const result = await audioRecorderPlayerAttachment.startRecorder(
-        path,
-        audioSet
-      );
-      setIsVoiceNoteRecording(true);
-      setVoiceNotesLink(result);
-      audioRecorderPlayerAttachment.addRecordBackListener((e) => {
-        const seconds = Math.floor(e.currentPosition / 1000);
-        if (seconds >= 900) {
-          setStopRecording(!stopRecording);
-        }
-        setVoiceNotes({
-          recordSecs: e.currentPosition,
-          recordTime: audioRecorderPlayerAttachment
-            .mmssss(Math.floor(e.currentPosition))
-            .slice(0, 5),
-          name: name,
-        });
-        return;
-      });
-    }
-  };
-
-  // to stop audio recording
-  const stopRecord = async () => {
-    if (isVoiceNoteRecording) {
-      await audioRecorderPlayerAttachment.stopRecorder();
-      audioRecorderPlayerAttachment.removeRecordBackListener();
-
-      // if isVoiceResult is true we show audio player instead of audio recorder
-      const voiceNote = {
-        uri: voiceNotesLink,
-        type: VOICE_NOTE_TEXT,
-        name: `${voiceNotes.name}.${isIOS ? "m4a" : "mp3"}`,
-        duration: Math.floor(voiceNotes.recordSecs / 1000),
-      };
-      dispatch({
-        type: SELECTED_VOICE_NOTE_FILES_TO_UPLOAD,
-        body: {
-          audio: [voiceNote],
-        },
-      });
-
-      setIsVoiceNoteRecording(false);
-
-      LMChatAnalytics.track(
-        Events.VOICE_NOTE_RECORDED,
-        new Map<string, string>([
-          [Keys.CHATROOM_TYPE, chatroomType?.toString()],
-          [Keys.CHATROOM_ID, chatroomID?.toString()],
-        ])
-      );
-    }
-  };
-
-  const handleStopRecord = async () => {
-    // to give some time for initiating the start recorder, then only stop it
-    setTimeout(async () => {
-      await stopRecord();
-      setIsVoiceResult(true);
-      setIsRecordingLocked(false);
-    }, 500);
-  };
-
-  // to reset all the recording data we had previously
-  const clearVoiceRecord = async () => {
-    if (isVoiceNoteRecording) {
-      await stopRecord();
-    } else if (isVoiceNotePlaying) {
-      await stopPlay();
-    }
-    setVoiceNotes({
-      recordSecs: 0,
-      recordTime: "",
-      name: "",
-    });
-    setVoiceNotesPlayer({
-      currentPositionSec: 0,
-      currentDurationSec: 0,
-      playTime: "",
-      duration: "",
-    });
-    setVoiceNotesLink("");
-    setIsRecordingLocked(false);
-
-    dispatch({
-      type: CLEAR_SELECTED_VOICE_NOTE_FILES_TO_UPLOAD,
-    });
-
-    // if isVoiceResult is false we show audio recorder instead of audio player
-    setIsVoiceResult(false);
-
-    LMChatAnalytics.track(
-      Events.VOICE_NOTE_CANCELED,
-      new Map<string, string>([
-        [Keys.CHATROOM_TYPE, chatroomType?.toString()],
-        [Keys.CHATROOM_ID, chatroomID?.toString()],
-      ])
-    );
-  };
-
-  // to start playing audio recording
-  const startPlay = async (path: string) => {
-    await audioRecorderPlayerAttachment.startPlayer(path);
-    audioRecorderPlayerAttachment.addPlayBackListener((e) => {
-      const playTime = audioRecorderPlayerAttachment.mmssss(
-        Math.floor(e.currentPosition)
-      );
-      const duration = audioRecorderPlayerAttachment.mmssss(
-        Math.floor(e.duration)
-      );
-      setVoiceNotesPlayer({
-        currentPositionSec: e.currentPosition,
-        currentDurationSec: e.duration,
-        playTime: audioRecorderPlayerAttachment
-          .mmssss(Math.floor(e.currentPosition))
-          .slice(0, 5),
-        duration: audioRecorderPlayerAttachment
-          .mmssss(Math.floor(e.duration))
-          .slice(0, 5),
-      });
-
-      // to reset the player after audio player completed it duration
-      if (playTime === duration) {
-        setIsVoiceNotePlaying(false);
-        setVoiceNotesPlayer({
-          currentPositionSec: 0,
-          currentDurationSec: 0,
-          playTime: "",
-          duration: "",
-        });
-      }
-      return;
-    });
-
-    LMChatAnalytics.track(
-      Events.VOICE_NOTE_PREVIEWED,
-      new Map<string, string>([
-        [Keys.CHATROOM_TYPE, chatroomType?.toString()],
-        [Keys.CHATROOM_ID, chatroomID?.toString()],
-      ])
-    );
-    setIsVoiceNotePlaying(true);
-  };
-
-  // to stop playing audio recording
-  const stopPlay = async () => {
-    await audioRecorderPlayerAttachment.stopPlayer();
-    setIsVoiceNotePlaying(false);
-  };
-
-  // to pause playing audio recording
-  const onPausePlay = async () => {
-    await audioRecorderPlayerAttachment.pausePlayer();
-    setIsVoiceNotePlaying(false);
-  };
-
-  // to resume playing audio recording
-  const onResumePlay = async () => {
-    await audioRecorderPlayerAttachment.resumePlayer();
-    setIsVoiceNotePlaying(true);
-  };
-
   const askPermission = async () => {
     if (!isIOS) {
       const permission = await requestAudioRecordPermission();
@@ -1998,13 +1565,6 @@ export const InputBoxContextProvider = ({
   const contextValues: InputBoxContext = {
     isVoiceNoteIconPress,
     hideDMSentAlert,
-    lockAnimatedStyles,
-    upChevronAnimatedStyles,
-    longPressGesture,
-    onUpdatePanGesture,
-    onEndPanGesture,
-    panGesture,
-    panStyle,
     handleVideoThumbnail,
     handleImageAndVideoUpload,
     selectGallery,
@@ -2025,14 +1585,6 @@ export const InputBoxContextProvider = ({
     detectLinkPreview,
     handleInputChange,
     onEdit,
-    startRecord,
-    stopRecord,
-    handleStopRecord,
-    clearVoiceRecord,
-    startPlay,
-    stopPlay,
-    onPausePlay,
-    onResumePlay,
     askPermission,
     setModalVisible,
     chatroomType,
@@ -2051,7 +1603,6 @@ export const InputBoxContextProvider = ({
     groupTags,
     taggedUserName,
     setMessage,
-    setFormattedConversation,
     setUserTaggingList,
     setGroupTags,
     setIsUserTagging,
@@ -2066,28 +1617,22 @@ export const InputBoxContextProvider = ({
     setIsEditable,
     isDoc,
     isGif,
-    isDeleteAnimation,
-    voiceNotes,
-    isVoiceResult,
-    micIconOpacity,
-    isRecordingLocked,
-    isVoiceNotePlaying,
-    voiceNotesPlayer,
-    voiceNotesLink,
     chatRequestState,
     inputHeight,
     setInputHeight,
     myRef,
     MAX_LENGTH,
     isPrivateMember,
-    isVoiceNoteRecording,
     isRecordingPermission,
-    composedGesture,
     setIsVoiceNoteIconPress,
     modalVisible,
     navigation,
     chatroomID,
     conversations,
+    isRecordingLocked,
+    setIsRecordingLocked,
+    voiceNotes,
+    setVoiceNotes,
   };
 
   return (
